face-encoder/requirements.txt
Keras==2.3.1
tensorflow-gpu==1.14.0
grpcio==1.27.2
pillow
